### 2022-06-27 #4520 @editorialbot
[PRE REVIEW]: River: a cross-platform, high-throughput, structured streaming framework
**Submitting author:** @pbotros (<a href="http://orcid.org/0000-0002-4797-1651">Paul Botros</a>)
**Repository:** https://github.com/pbotros/river
**Branch with paper.md** (empty if default branch): botros/manuscript
**Version:** v1.0.1
