### 2022-06-29 #4520 @Kevin-Mattheus-Moerman
@pbotros please amend your paper to include a _Statement of need section_ as per [our guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). You can call `@editorialbot generate pdf` here to update the paper when ready.
