### 2022-07-04 #4520 @jedbrown
Sorry about my slow reply. I suppose I could, but it's pretty outside my area. I think the paper would benefit greatly from a cartoon showing an example deployment with producers, consumers, representative data bandwidth, and associated stakeholders (if there are multiple involved in a single deployment).
