+++ manuscript/paper.md
@@ -27,9 +27,11 @@ Modern experiments in neuroscience demand increasingly complex and performant sy
 
 # Summary
 
-To tackle these problems, we developed River, a structured streaming solution to streamline data management for modern neuroscience experiments. Producers of data – “writers” – first define a structure for each data sample for a particular stream. Then, utilizing function calls similar to those of file I/O, writers can then write data to that stream, while an unlimited number of “readers” can consume this data in real-time by specifying the appropriate stream name. Each reader can read at its own pace, with support for blocking for a given period of time or number of samples. Readers and writers only require network connectivity to a centralized server, enabling a horizontal scaling paradigm useful for computationally-intensive or cloud-reliant experiments. Finally, leveraging the structured nature of each stream, a separate “ingester” process consumes each stream’s data and writes data to disk in a cross-platform tabular format, [Parquet](https://github.com/apache/parquet-format), for post-hoc analysis. Once persisted by the ingester, sufficiently stale data is then deleted from the River stream, enabling streams to be indefinitely long. River is written in C++ with Python and MATLAB bindings and works under Linux, Mac OSX, and Windows.
+To tackle these problems, we developed River, a structured streaming solution to streamline data management for modern neuroscience experiments. Producers of data – “writers” – first define a structure for each data sample for a particular stream. Then, utilizing function calls similar to those of file I/O, writers can then write data to that stream, while an unlimited number of “readers” can consume this data in real-time by specifying the appropriate stream name. Each reader can read at its own pace, with support for blocking for a given period of time or number of samples. Readers and writers only require network connectivity to a centralized server, enabling a horizontal scaling paradigm useful for computationally-intensive or cloud-reliant experiments. Finally, leveraging the structured nature of each stream, a separate “ingester” process consumes each stream’s data and writes data to disk in a cross-platform tabular format, [Parquet](https://github.com/apache/parquet-format), for post-hoc analysis. Once persisted by the ingester, sufficiently stale data is then deleted from the River stream, enabling streams to be indefinitely long. River is written in C++ with Python and MATLAB bindings and works under Linux, Mac OSX, and Windows. An example deployment, highlighting River architecture, can be found in \autoref{fig:deployment}.
 
-Eschewing a bespoke, research-specific solution to ease its maintenance burden, we utilized an industry-standard database, [Redis](https://redis.io), to underlie River, enabling sub-millisecond latencies, low jitter, and 20+ MB/s throughput. Critical to researchers, River guarantees that no written data can be dropped by any readers, even in the case of long pauses in reading, and that all persisted data exactly mirrors the data streamed online. These guarantees come in contrast to the best-effort paradigm of another popular research-oriented streaming framework [@labstreaminglayer]. Additionally, timestamp management is centralized and thus stream synchronization is simplified. Other streaming solutions widely used in the software industry, such as [Kafka](https://kafka.apache.org/) and [RabbitMQ](https://www.rabbitmq.com/), have more strict guarantees on data availability and delivery but subsequently suffer in raw streaming performance and can be difficult to set up robustly.
+![Example deployment of 1 writer and 2 readers utilizing River, where the writer reads data from a probe, one reader visualizes the data, and another reader utilizes the data to actuate an external device. Readers and writers may exist on different devices, with network connectivity to the in-memory database, Redis, the only requirement. Dotted line demarcates the boundary of River APIs, with monospaced-font commands referring to River APIs invoked to write, read, and ingest data. Note all interactions with the underlying database, Redis, and nuances of ingestion are abstracted behind River APIs, enabling users to both stream online and analyze offline data with simpler commands.\label{fig:deployment}](deployment-01.png)
+
[…]
