### 2022-08-11 #4627 @agricolab
- [ ] Tests: The authors claim that their package is ["well-tested via both unit and integration tests"](https://pbotros.github.io/river/docs/contributing.html#contributing), but i failed to find code for the tests in the repo, and could not find a documentation on how to run the tests myself.
- [ ] Running the examples and installing the required dependencies resulted in an binary incompatibility issue that prevents me from further testing the package. See https://github.com/pbotros/river/issues/20
- [ ] Installation dependencies: Authors might want to clarify upfront that it requires Redis, and especially with which versions it is compatible.
- [ ] Contribution guidelines are very short and would benefit from expansion (i.e. how are pull requests handled, e.g. always create an issue first, or just fork and PR, etc...)  and possibly, an inclusion of a Code of Conduct.
- [ ] The online documentation seems to miss a statement of need
- [ ] The paper lacks a state of the field comparison. Authors cite LSL as the dominant multi-modal data stream and receive package used in neuroscience. Compared to LSL, their solution is very heavy-weight and has many dependencies. The submission could benefit from clarifying advantages and disadvantages. Additionally, they might consider to include a short rationale why they choose Redis as a backend.
