### 2022-09-26 #4627 @ixjlyons
## General comments

It's mentioned in the paper and documentation that the ingester will remove stale data from the Redis DB once it's deemed sufficiently stale -- it's not quite clear to me how staleness is determined. Perhaps some additional detail could be added to the tutorial.

While I did find that a `river_test` binary is built, I didn't see anything in the documentation referring to it for running the tests (the tutorial gives an example command for running `river_benchmark`. Perhaps a note about the tests could go in the Contributing page. The Contributing page could also use a note about getting support -- should a user create an issue or reach out via some other means to get help with a problem?

Regarding performance verification: I was able to run the performance notebook on an aging laptop with a few minor changes and confirmed 20MB/s+ throughput and sub-millisecond latency as claimed. However, this was with Redis running locally.

## Docs

### Installation

Pre-packaged installation with conda on Linux worked as expected.

A few notes from attempting a source build installation on OpenSUSE Tumbleweed (eventually succeeded):

- Build instructions say `mkdir build/release` then `cmake ../..`, but the `CMakeLists.txt` is within `cpp/`
- Missing parquet-cpp in dependency list for ingester
- I had to set `-Wno-maybe-uninitialized`, apparently for glog itself. May be a gcc thing (gcc version 12.1.1)
- There doesn't seem to be any instruction for building the Python bindings

### Tutorial

- It might be nice to limit the line length in the example code to avoid needing to scroll horizontally
- Malformed link to "Redis website"

## Paper

- Line 22: may want to define "I/O"
- Line 32: suggestion: "...highlighting River's architecture..." or "...highlighting the River architecture..."
