### 2022-12-22 c7bf877
Adding INT16 support, atomic metadata write, and build changes
- Adding INT16 support across C++,Python,MATLAB. Adds INT16
support similar to that of INT32/INT64 that already exists.
- Adding atomic write of user metadata on initialization of stream.
Before, we were writing the user metadata right after the internal
metadata field in Redis, meaning sometimes readers could see no user
metadata even after initialization. This unifies those, since setting a
hash in Redis is atomic.
- Latest GLOG (at least on building Mac ARM) throws an error when
  attempting to format a long or int in a string. 
[…]
