+++ cpp/CMakeLists.txt
@@ -26,7 +26,7 @@ if (RIVER_BUILD_TESTS)
   FetchContent_Declare(
     googletest
     GIT_REPOSITORY    https://github.com/google/googletest.git
-    GIT_TAG           release-1.10.0
+    GIT_TAG           release-1.12.1
   )
   FetchContent_MakeAvailable(googletest)
   enable_testing()

[…]
