+++ python/setup.py.in
@@ -25,7 +25,7 @@ setup(
             include_dirs=filter_empty([
                 np.get_include(),
             ] + '${RIVER_INCLUDE_DIRS}'.split(';')),
-            extra_compile_args=["-std=c++14"]),
+            extra_compile_args=["-std=c++17"]),
         compiler_directives={
             'language_level' : "3",
             "embedsignature": True,
