### 2023-01-04 ecb3c51
adding river install benchmark
