### 2023-01-04 aeb420f
adding river install benchmark
