### 2023-01-19 e908261
Updating docs with new Sphinx & a few corrections
Lots of small changes from upgrading Sphinx and its theme, but otherwise
correcting & adding information on both installation and contributing.
