### 2023-01-26 430f79f
Fixed bug and added test for INT16 support
I only partially / incorrectly added INT16 support in a previous commit.
This fixes it.
