### 2023-01-27 bed4be7
Computing local vs global server clocks optionally (& other fixes)
Defaulting to not computing this. Only if needed to calibrate local and
server times, i.e. if you're running a writer on a different node than
the others and there's no NTP/chrony sync syncing clocks.

Other fixes:
- Allowing a timeout in seconds to be specified within RedisConnection
  optionally
- Not freeing reply pointers that are NULL
