### 2023-01-28 02eb9ac
Allowing further customization of ingester
Enabling configuration at a finer level for ingestion of streams. Allows
a JSON settings file to be passed on the command line that can give
finer-grained control of streams via regexes.

See example_settings/example.json for an example JSON setting file.
Defaults to settings that include all streams and has default settings.
