+++ cpp/ingester/src/ingester_settings.h
@@ -1,0 +1,146 @@
+//
+// Created by Paul Botros on 11/15/19.
+//
+
+#ifndef PARENT_INGESTER_SETTINGS_H
+#define PARENT_INGESTER_SETTINGS_H
+
+
+#include <vector>
+#include <optional>
+#include <regex>
+#include <algorithm>
+#include <river.h>
+#include <nlohmann/json.hpp>
+
+using json = nlohmann::json;
+
+static const int64_t DEFAULT_BYTES_PER_ROW_GROUP = 134217728LL; // 128MB
+static const int DEFAULT_MINIMUM_AGE_SECONDS_BEFORE_DELETION = 60; // 60s lookback
+
+class StreamIngestionSettings {
+public:
+    StreamIngestionSettings() {}
+
+    StreamIngestionSettings(
+            std::optional<std::vector<std::regex>> columns_blacklist_,
+            std::optional<std::vector<std::regex>> columns_whitelist_,
+            int64_t bytes_per_row_group_,
+            int minimum_age_seconds_before_deletion_) {
+        // Whitelist takes priority over blacklist; see filter below.
+        columns_blacklist = columns_blacklist_;
+        columns_whitelist = columns_whitelist_;
+        bytes_per_row_group = bytes_per_row_group_;
+        minimum_age_seconds_before_deletion = minimum_age_seconds_before_deletion_;
+    }
+
+    std::vector<river::FieldDefinition> Filter(const std::vector<river::FieldDefinition>& fields) const {
+        if (columns_whitelist.has_value()) {
+            return FilterList(fields, columns_whitelist.value(), false);
+        } else if (columns_blacklist.has_value()) {
+            return FilterList(fields, columns_blacklist.value(), true);
+        } else {
+            std::vector<river::FieldDefinition> ret(fields.begin(), fields.end());
+            return ret;
+        }
+    }
+
+    std::optional<std::vector<std::regex>> columns_blacklist;
+    std::optional<std::vector<std::regex>> columns_whitelist;
+    int64_t bytes_per_row_group;
+    int minimum_age_seconds_before_deletion;
+
+
+private:
+    static std::vector<river::FieldDefinition> FilterList(
+            const std::vector<river::FieldDefinition>& fields,
+            const std::vector<std::regex>& list,
+            bool default_value) {
+        std::vector<river::FieldDefinition> ret;
+        for (const auto& field : fields) {
+            bool should_add = default_value;
+            for (const auto &column_whitelist: list) {
+                if (std::regex_match(field.name, column_whitelist)) {
+                    should_add = !should_add;
+                    break;
+                }
+            }
+            if (should_add) {
+                ret.push_back(field);
+            }
+        }
+        return ret;
+    }
+};
+
+static std::vector<std::pair<std::regex, StreamIngestionSettings>> ParseStreamSettingsJson(const json& settings_json) {
+    int64_t bytes_per_row_group_global = DEFAULT_BYTES_PER_ROW_GROUP;
+    int minimum_age_seconds_before_deletion_global = DEFAULT_MINIMUM_AGE_SECONDS_BEFORE_DELETION;
+    if (settings_json.contains("global_settings")) {
+        minimum_age_seconds_before_deletion_global = settings_json.value(
+                "minimum_age_seconds_before_deletion",
+                minimum_age_seconds_before_deletion_global);
+        bytes_per_row_group_global = settings_json.value("bytes_per_row_group", bytes_per_row_group_global);
+    }
+
+    std::vector<std::pair<std::regex, StreamIngestionSettings>> ret;
+    if (!settings_json.contains("stream_settings")) {
+        LOG(WARNING) << "Warning: stream settings was empty. Was that intentional to not consume any streams?"
+                     << std::endl;
+        return ret;
+    }
+
[…]
