### 2023-01-28 c9c8150
Fixing the build on Linux and killing a warning
