### 2023-04-30 aaf0250
Adding samples_per_read as a config
