### 2023-05-08 3d876e8
Not using high_resolution_clock
Per the C++ standard documentation, high_resolution_clock isn't
recommended; instead, steady_clock or system_clock should be used when
possible.
