### 2023-08-10 abf3ea4
Improving throughput in stream writing
- introduces a Redis server module, "river" (built in the redismodule
  subdirectory) that can be optionally installed into Redis to provide
  much improved throughput. It minimizes the number of bytes transmitted
  along the network by combining all existing XADD calls into a single
  call.
- If the module is installed, we also carefully avoid copies of the data
  to be sent, by manually managing the formatting of the Redis command
  and the sending via sockets.
