+++ cpp/src/redismodule/river_redismodule.c
@@ -1,0 +1,127 @@
+
+#include <redismodule.h>
+
+#include <rmutil/util.h>
+#include <rmutil/strings.h>
+#include <rmutil/test_util.h>
+
+#define ASSERT_NOERROR_STRING_FXN(r) \
+  if ((r) != REDISMODULE_OK) { \
+    return REDISMODULE_ERR; \
+  }
+
+int BatchXaddCommand(RedisModuleCtx *ctx, RedisModuleString **argv, int argc) {
+    if (argc != 6) {
+        return RedisModule_WrongArity(ctx);
+    }
+    RedisModule_AutoMemory(ctx);
+
+    // open the key and make sure it's indeed a STREAM
+    RedisModuleKey *key =
+        RedisModule_OpenKey(ctx, argv[1], REDISMODULE_READ | REDISMODULE_WRITE);
+    if (RedisModule_KeyType(key) != REDISMODULE_KEYTYPE_STREAM &&
+        RedisModule_KeyType(key) != REDISMODULE_KEYTYPE_EMPTY) {
+        return RedisModule_ReplyWithError(ctx, REDISMODULE_ERRORMSG_WRONGTYPE);
+    }
+
+    long long index_start;
+    long long num_samples;
+    long long sample_size_bytes;
+    ASSERT_NOERROR_STRING_FXN(RedisModule_StringToLongLong(argv[2], &index_start));
+    ASSERT_NOERROR_STRING_FXN(RedisModule_StringToLongLong(argv[3], &num_samples));
+    ASSERT_NOERROR_STRING_FXN(RedisModule_StringToLongLong(argv[4], &sample_size_bytes));
+
+    size_t value_length;
+    const char *value = RedisModule_StringPtrLen(argv[5], &value_length);
+
+    RedisModuleString **xadd_params = (RedisModuleString **) RedisModule_Alloc(sizeof(RedisModuleString *) * 4);
+    for (long long i = 0; i < num_samples; i++) {
+        const char *i_str = "i";
+        xadd_params[0] = RedisModule_CreateString(ctx, i_str, strlen(i_str));
+        xadd_params[1] = RedisModule_CreateStringFromLongLong(ctx, index_start + i);
+
+        const char *val_str = "val";
+        xadd_params[2] = RedisModule_CreateString(ctx, val_str, strlen(val_str));
+
+        size_t sample_start = i * sample_size_bytes;
+        xadd_params[3] = RedisModule_CreateString(ctx, &value[sample_start], sample_size_bytes);
+
+        // fields and values;
+        int stream_add_resp = RedisModule_StreamAdd(key, REDISMODULE_STREAM_ADD_AUTOID, NULL, xadd_params, 2);
+        if (stream_add_resp != REDISMODULE_OK) {
+            return RedisModule_ReplyWithError(ctx, "ERR Xadd failed.");
+        }
+
+        for (int j = 0; j < 4; j++) {
+            RedisModule_FreeString(ctx, xadd_params[j]);
+        }
+     }
+    RedisModule_Free(xadd_params);
+    return RedisModule_ReplyWithSimpleString(ctx, "OK");
+}
+
+int BatchXaddVariableCommand(RedisModuleCtx *ctx, RedisModuleString **argv, int argc) {
+    // batch_xadd_variable <key> <index start> <sizes in ints> <value in bytes>
+    if (argc != 5) {
+        return RedisModule_WrongArity(ctx);
+    }
+    RedisModule_AutoMemory(ctx);
+
+    // open the key and make sure it's indeed a STREAM
+    RedisModuleKey *key =
+        RedisModule_OpenKey(ctx, argv[1], REDISMODULE_READ | REDISMODULE_WRITE);
[…]
