### 2023-08-11 #4627 @pbotros
@gkthiruvathukal sorry for the delay here. I do still want to pick this back up but it seems there always are competing priorities. We're making some improvements to the codebase now so hopefully can include better metrics, comparisons, & address feedback here too.

Would it make sense to close this for now, and once I am ready & have addressed comments, I can open this review back up and pick back up where we left off? Thanks!
