+++ cpp/src/redis.cpp
@@ -10,6 +10,11 @@
 #include <glog/logging.h>
 #include <fmt/format.h>
 
+#include "net/sockcompat.h"
+#if defined(_WIN32) || defined(_WIN64)
+#include <windows.h>
+#endif
+
 using json = nlohmann::json;
 using namespace std;
 
@@ -379,5 +384,45 @@ void Redis::DeleteMetadata(const string &stream_name) {
         throw RedisException(msg);
     }
 }
+
+void __redisSetError(redisContext *c, int type, const char *str) {
+    // Lifted directly from Redis.
+    size_t len;
+
+    c->err = type;
+    if (str != NULL) {
+        len = strlen(str);
+        len = len < (sizeof(c->errstr)-1) ? len : (sizeof(c->errstr)-1);
+        memcpy(c->errstr,str,len);
+        c->errstr[len] = '\0';
+    } else {
+        /* Only REDIS_ERR_IO may lack a description! */
+        assert(type == REDIS_ERR_IO);
+#ifdef _WIN32
+#define strerror_r(errno,buf,len) strerror_s(buf,len,errno)
+#endif /* _WIN32 */
+        strerror_r(errno, c->errstr, sizeof(c->errstr));
+    }
+}
+
+int Redis::SendCommandPreformatted(std::vector<std::pair<const char *, size_t>> preformatted_commands) {
+    int nwritten_total = 0;
+    for (int i = 0; i < preformatted_commands.size(); i++) {
+        int nwritten = redis_sockcompat::send_redis_sockcompat(
+            _context->fd, preformatted_commands[i].first, preformatted_commands[i].second, 0);
+        if (nwritten < 0) {
+            if ((errno == EWOULDBLOCK && !(_context->flags & REDIS_BLOCK)) || (errno == EINTR)) {
+                /* Try again later */
+            } else {
+                __redisSetError(_context, REDIS_ERR_IO, NULL);
+                return -1;
+            }
+        }
+        nwritten_total += nwritten;
+    }
+    return nwritten_total;
+}
+
+
 }
 }
