### 2023-08-11 c1e4828
Implementing optional compression for river streams
Supporting ZFP-based compression/decompression that's transparent to the
user of the River streams.
