+++ cpp/src/compression/compressor.cpp
@@ -1,0 +1,248 @@
+//
+// Created by Paul Botros on 8/11/23.
+//
+
+#include "compressor.h"
+#include <zfp.hpp>
+#include <sstream>
+
+
+namespace river {
+
+class ZfpCompressorImpl {
+public:
+    int num_rows_;
+    int num_cols_;
+    std::vector<char> buffer_;
+    zfp_stream *zfp_;
+    zfp_field  *field_;
+
+    ~ZfpCompressorImpl() {
+        if (zfp_) {
+            if (zfp_->stream) {
+                stream_close(zfp_->stream);
+            }
+            zfp_stream_close(zfp_);
+        }
+        if (field_) {
+            zfp_field_free(field_);
+        }
+    }
+};
+
+template <class DataTypeT>
+constexpr zfp_type zfp_type_for_class() {
+    if constexpr (std::is_same_v<DataTypeT, int16_t> || std::is_same_v<DataTypeT, int32_t>) {
+        return zfp_type_int32;
+    } else if constexpr (std::is_same_v<DataTypeT, float>) {
+        return zfp_type_float;
+    } else if constexpr (std::is_same_v<DataTypeT, double>) {
+        return zfp_type_double;
+    } else {
+        throw std::invalid_argument("Unhandled data type for conversion to ZFP");
+    }
+}
+
+template <class DataTypeT>
+ZfpCompressor<DataTypeT>::ZfpCompressor(int num_cols, double tolerance)
+    : impl_(nullptr) {
+    this->num_cols_ = num_cols;
+    this->tolerance_ = tolerance;
+}
+
+template <class DataTypeT>
+ZfpCompressor<DataTypeT>::~ZfpCompressor() noexcept {
+    if (impl_ != nullptr) {
+        delete impl_;
+        impl_ = nullptr;
+    }
+}
+
+template <class DataTypeT>
+std::vector<char> ZfpCompressor<DataTypeT>::compress(const char *data, size_t length) {
+    // assert(length % (sizeof(DataTypeT) * num_cols_) == 0);
+    auto num_rows = (int) (length / sizeof(DataTypeT) / num_cols_);
+
+    if (impl_ == nullptr || impl_->num_rows_ != num_rows) {
+        // We lazily instantiate the internal data structure within this compressor to minimize the number of
+        // initializations we do. This paradigm leverages the fact that most StreamWriter's will likely call Write()
+        // with the same number of rows of data every time.
+
+        // Clean up past impl_ struct (delete nullptr is okay)
+        delete impl_;
+
+        // initialize metadata for a compressed stream
+        impl_ = new ZfpCompressorImpl();
+        impl_->num_rows_ = num_rows;
+        impl_->num_cols_ = this->num_cols_;
+        impl_->zfp_ = zfp_stream_open(nullptr);
+
+        impl_->field_ = zfp_field_2d(
+            nullptr,
+            zfp_type_for_class<DataTypeT>(),
+            impl_->num_cols_,
+            impl_->num_rows_);
+
+        size_t bufsize = zfp_stream_maximum_size(impl_->zfp_, impl_->field_);
+        impl_->buffer_.resize(bufsize);
+
+        // initialize metadata for a compressed stream
+        if (tolerance_ <= 0.0) {
+            zfp_stream_set_reversible(impl_->zfp_);
+        } else {
+            zfp_stream_set_accuracy(impl_->zfp_, tolerance_);
+        }
+
+        // associate bit stream with allocated buffer
+        bitstream *stream = stream_open(impl_->buffer_.data(), impl_->buffer_.size());
+        zfp_stream_set_bit_stream(impl_->zfp_, stream);
+    }
+
[…]
