### 2023-08-17 6ebaf83
Bug fixes for tail, seek for compressed flows
- Tail and seek weren't properly implemented for compressed flows
- Data promotion was incorrect, resulting in segfaults
- Adding DUMMY compressor for testing
- More exception handling in the StreamReader()
- Better integration tests to cover tail/seek
