+++ cpp/src/reader.cpp
@@ -183,7 +183,29 @@ int64_t StreamReader::ReadBytes(
             int len;
             const char *value = FindField(element, "val", &len);
 
-            if (value != nullptr) {
+            if (value == nullptr) {
+                if (!this->decompressor_) {
+                    continue;
+                }
+
+                if (lookahead_data_cache_index_ > (lookahead_data_cache_.size() - sample_size_)) {
+                    if (FindField(element, "reference", nullptr) == nullptr) {
+                        // This is a non-value field (like an EOF or tombstone), so skip.
+                        continue;
+                    }
+
+                    // We are compressed, but we're out of elements. Should never happen, but at least prevent a
+                    // SEGFAULT by explicitly throwing an exception.
+                    throw StreamReaderException("Lookahead data cache empty, but expected an element.");
+                }
+
+                memcpy(
+                    &buffer[buffer_index],
+                    lookahead_data_cache_.data() + lookahead_data_cache_index_,
+                    sample_size_);
+                lookahead_data_cache_index_ += sample_size_;
+                buffer_index += sample_size_;
+            } else {
                 if (sizes != nullptr) {
                     (*sizes)[samples_fetched] = len;
                 }
@@ -192,12 +214,14 @@ int64_t StreamReader::ReadBytes(
                 }
 
                 if (this->decompressor_) {
+                    // Compressed stream but we're on an element with "val", meaning we need to repopulate our cache
+                    // and extract out the relevant element we're on.
                     lookahead_data_cache_ = decompressor_->decompress(value, len);
-                    lookahead_data_cache_index_ = 0;
-
-                    memcpy(&buffer[buffer_index], lookahead_data_cache_.data() + lookahead_data_cache_index_, sample_size_);
-                    lookahead_data_cache_index_ += sample_size_;
-
+                    memcpy(
+                        &buffer[buffer_index],
+                        lookahead_data_cache_.data(),
+                        sample_size_);
+                    lookahead_data_cache_index_ = sample_size_;
                     buffer_index += sample_size_;
                 } else if (this->has_variable_width_field_) {
                     memcpy(&buffer[buffer_index], value, len);
@@ -206,16 +230,9 @@ int64_t StreamReader::ReadBytes(
                     memcpy(&buffer[buffer_index], value, sample_size_);
                     buffer_index += sample_size_;
                 }
-                samples_fetched++;
-                num_samples_read_++;
-            } else if (this->decompressor_ && lookahead_data_cache_index_ < lookahead_data_cache_.size()) {
-                memcpy(&buffer[buffer_index], lookahead_data_cache_.data() + lookahead_data_cache_index_, sample_size_);
-                lookahead_data_cache_index_ += sample_size_;
-                buffer_index += sample_size_;
-
-                samples_fetched++;
-                num_samples_read_++;
             }
+            samples_fetched++;
+            num_samples_read_++;
         }
 
         redisReply *last_element = data_reply->element[data_reply->elements - 1];
@@ -275,6 +292,51 @@ int64_t StreamReader::ReadBytes(
     }
 
   return samples_fetched;
+}
+
+void StreamReader::ReloadLookaheadCache(const char *val_str, int val_str_len, const redisReply *values) {
+    if (!decompressor_) {
+        return;
+    }
+
+    int64_t decompressed_sample_offset;
+    if (val_str != nullptr) {
+        // We got to the data sample that contains the encrypted value, so load it directly
+        decompressed_sample_offset = 0;
+        lookahead_data_cache_ = decompressor_->decompress(val_str, val_str_len);
+    } else {
+        // We got a data sample that follows a compressed blob. There should be a "reference" key that
+        // then references the source of truth
+        int reference_str_len;
+        const char *reference_str = FindField(values, "reference", &reference_str_len);
+        if (reference_str == nullptr) {
[…]
