### 2023-08-17 2e5de0c
Adding RIVER_INSTALL_BENCHMARK flag
