### 2023-08-17 c3eb59e
Minor build and benchmark fixes
