### 2023-09-14 a515b64
Truly making ZFP optional
It appears that Mac ARM doesn't have the appropriate ZFP binaries
compiled for ARM. Regardless, making the RIVER_BUILD_ZFP flag actually
work properly and not require installation of the ZFP library.
