### 2023-09-18 be59133
Adding OWNER_EXECUTE to redis module
