### 2023-09-18 c51292e
Exposing an HTTP server for the ingester
Allows dynamic querying of streams as well as direct retrieval of
metadata and data contents.
