+++ cpp/ingester/src/ingester_http_server.cpp
@@ -1,0 +1,176 @@
+//
+// Created by Paul Botros on 9/18/23.
+//
+
+#include "ingester_http_server.h"
+#include <set>
+#include <optional>
+#include <filesystem>
+#include <utility>
+#include <glog/logging.h>
+
+using json = nlohmann::json;
+
+namespace river {
+
+IngesterHttpServer::IngesterHttpServer(std::filesystem::path root_directory, int port) :
+    root_directory_(std::move(root_directory)), port_(port) {}
+
+void IngesterHttpServer::Start() {
+    svr_ = std::make_unique<httplib::Server>();
+    svr_running_thread_ = std::make_unique<std::thread>(&IngesterHttpServer::RunServerForever, this);
+}
+
+void IngesterHttpServer::Stop() {
+    if (svr_) {
+        svr_->stop();
+    }
+    if (svr_running_thread_) {
+        if (svr_running_thread_->joinable()) {
+            svr_running_thread_->join();
+        }
+        svr_running_thread_.reset();
+    }
+    svr_.reset();
+}
+
+void IngesterHttpServer::RunServerForever() {
+    svr_->Get("/api/streams", [this](const httplib::Request &req, httplib::Response &res) {
+        this->HandleStreamsGet(req, res);
+    });
+    svr_->Get(R"(/api/streams/([^/]+))", [this](const httplib::Request &req, httplib::Response &res) {
+        auto stream_name = req.matches[1];
+        this->HandleStreamGet(req, res, stream_name);
+    });
+    svr_->Get(R"(/api/streams/([^/]+)/data.parquet)", [this](const httplib::Request &req, httplib::Response &res) {
+        auto stream_name = req.matches[1];
+        this->HandleStreamGetParquet(req, res, stream_name);
+    });
+
+    svr_->listen("0.0.0.0", port_);
+}
+
+void IngesterHttpServer::HandleStreamsGet(const httplib::Request &request, httplib::Response &response) {
+    //--- filenames are unique so we can use a set
+    std::set<std::filesystem::path> sorted_by_name;
+
+    for (auto &entry : std::filesystem::directory_iterator(root_directory_)) {
+        sorted_by_name.insert(entry.path());
+    }
+
+    json streams_json;
+    for (const auto &stream_name : sorted_by_name) {
+        auto stream_json = GetStreamJson(stream_name);
+        if (stream_json) {
+            streams_json.push_back(stream_json.value());
+        }
+    }
+
+    json response_json;
+    response_json["streams"] = streams_json;
+    response.set_content(response_json.dump(), "application/json");
+}
+
+void IngesterHttpServer::HandleStreamGet(const httplib::Request &request,
+                                         httplib::Response &response,
+                                         const std::string &stream_name) {
+    json response_json;
+    auto val = GetStreamJson(stream_name);
+    if (!val.has_value()) {
+        response.set_content("Invalid stream name", "text/plain");
+        response.status = 404;
+        return;
+    }
+
+    response_json["stream"] = val.value();
+    response.set_content(response_json.dump(), "application/json");
+}
+
+void IngesterHttpServer::HandleStreamGetParquet(const httplib::Request &request,
+                                                httplib::Response &response,
+                                                const std::string &stream_name) {
+    auto data_filename = root_directory_ / stream_name / "data.parquet";
+    if (!std::filesystem::exists(data_filename)) {
+        response.set_content("Could not find data file", "text/plain");
+        response.status = 404;
+        return;
+    }
+
+    std::streampos data_f_size;
+    {
+        std::ifstream data_f(data_filename, std::ios::binary);
+        data_f_size = data_f.tellg();
+        data_f.seekg(0, std::ios::end);
[…]
