### 2023-09-19 8b59031
Not making +x for the redis module
Breaks conda-forge packaging. Before installation, the redis module must
be +x'd.
