### 2023-09-20 dbd10a7
Adding optional import and no SSL for httplib
